Template for InputDialog
