+++ frontend/src/components/dialogs/inputDialogs/inputDialogTemplateContent.js
@@ -1,0 +1,13 @@
+import {DialogContent} from "@mui/material";
+
+function InputDialogTemplateContent({
+    content
+                                    }) {
+    return (
+        <DialogContent dividers>
+            {content}
+        </DialogContent>
+    )
+}
+
+export default InputDialogTemplateContent;
